% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_uf.R
\docType{data}
\name{df_uf}
\alias{df_uf}
\title{Brazil state polygons}
\format{A data frame with 27 rows and 2 variables:
\describe{
   \item{NM_ESTADO}{State name}
   \item{NM_REGIAO}{Region name}
   \item{CD_GEOCUF}{UF code}
   \item{geometry}{multipolygon}
}}
\source{
IBGE.
}
\usage{
df_uf
}
\description{
An \code{sf} object to draw maps
}
\keyword{datasets}
